+++ src/arith_uint256.h
@@ -6,8 +6,6 @@
 #ifndef BITCOIN_ARITH_UINT256_H
 #define BITCOIN_ARITH_UINT256_H
 
-#include "blob_uint256.h"
-#include "uint512.h"
 #include <assert.h>
 #include <cstring>
 #include <stdexcept>
@@ -15,8 +13,6 @@
 #include <string>
 #include <vector>
 
-class blob_uint512;
-class blob_uint256;
 class uint256;
 class uint512;
 
@@ -363,6 +359,9 @@ public:
     arith_uint256& SetCompact(uint32_t nCompact, bool *pfNegative = NULL, bool *pfOverflow = NULL);
     uint32_t GetCompact(bool fNegative = false) const;
     uint32_t Get32(int n = 0) const { return pn[2 * n]; }
+
+    friend arith_uint256 UintToArith256(const uint256 &a);
+    friend uint256 ArithToUint256(const arith_uint256 &a);
 };
 
 /** 512-bit unsigned big integer. */
@@ -376,14 +375,15 @@ public:
 
     uint256 trim256() const;
 
-    //friend arith_uint512 UintToArith512(const blob_uint512 &a);
-    //friend blob_uint512 ArithToUint512(const arith_uint512 &a);
+    friend arith_uint512 UintToArith512(const uint512 &a);
+    friend uint512 ArithToUint512(const arith_uint512 &a);
 
 };
 
-/** Old classes definitions */
-
-/** End classes definitions */
+uint256 ArithToUint256(const arith_uint256 &);
+arith_uint256 UintToArith256(const uint256 &);
+uint512 ArithToUint512(const arith_uint512 &);
+arith_uint512 UintToArith512(const uint512 &);
 
 const arith_uint256 ARITH_UINT256_ZERO = arith_uint256();
 const arith_uint256 ARITH_UINT256_ONE = arith_uint256(1);
