+++ src/sapling/zip32.h
@@ -5,7 +5,7 @@
 #ifndef PIVX_ZIP32_H
 #define PIVX_ZIP32_H
 
-#include "blob_uint256.h"
+#include "uint256.h"
 #include "key.h"
 #include "sapling/address.h"
 #include "serialize.h"
